requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.13
django-database-url==1.0.3
gunicorn==20.1.0
PSycopg2==2.9.5
sqlparse==0.4.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
